move `HardFork` to `types`
