+++ consensus/rules/src/blocks.rs
@@ -5,12 +5,7 @@ use monero_serai::block::Block;
 
 use cuprate_cryptonight::*;
 
-use crate::{
-    current_unix_timestamp,
-    hard_forks::HardForkError,
-    miner_tx::{check_miner_tx, MinerTxError},
-    HardFork,
-};
+use crate::{current_unix_timestamp, hard_forks::HardForkError, miner_tx::{check_miner_tx, MinerTxError}, HardFork, check_block_version_vote};
 
 const BLOCK_SIZE_SANITY_LEEWAY: usize = 100;
 const BLOCK_FUTURE_TIME_LIMIT: u64 = 60 * 60 * 2;
@@ -249,11 +244,11 @@ pub fn check_block(
     block_blob_len: usize,
     block_chain_ctx: &ContextToVerifyBlock,
 ) -> Result<(HardFork, u64), BlockError> {
-    let (version, vote) = HardFork::from_block_header(&block.header)?;
+    let (version, vote) = HardFork::from_block_header(&block.header).map_err(|_| HardForkError::HardForkUnknown)?;
 
-    block_chain_ctx
-        .current_hf
-        .check_block_version_vote(&version, &vote)?;
+    check_block_version_vote(&block_chain_ctx
+        .current_hf,
+        &version, &vote)?;
 
     if let Some(median_timestamp) = block_chain_ctx.median_block_timestamp {
         check_timestamp(block, median_timestamp)?;
