+++ types/Cargo.toml
@@ -13,6 +13,7 @@ default    = ["blockchain", "epee", "serde"]
 blockchain = []
 epee       = ["dep:cuprate-epee-encoding"]
 serde      = ["dep:serde"]
+proptest  = ["dep:proptest-derive"]
 
 [dependencies]
 cuprate-epee-encoding = { path = "../net/epee-encoding", optional = true }
@@ -23,5 +24,8 @@ curve25519-dalek = { workspace = true }
 monero-serai     = { workspace = true }
 serde            = { workspace = true, features = ["derive"], optional = true }
 borsh            = { workspace = true, optional = true }
+thiserror        = { workspace = true }
+
+proptest-derive  = { workspace = true, optional = true }
 
 [dev-dependencies]
